Implements the checkMovement algorithm
This commit also adds some utility functions in the Coordinates class. I'm not
sure if these functions are well-suited for our board, because it's a grid but
the indexing is somewhat exotic. To be tested.
Also fixes a small bug were the origin of a moved pawn would not be erased.