More abstraction in AbstractBoard
The internal representation is not exposed anymore in AbstractBoard
and the actual implementation is left to subclasses.
Thus, every access to a Square must go through the abstract `getSquare()`
method, and the subclasses must redefine methods to access the width and
height of the board, as well as exposing a set of coordinates to access
the board's squares.