Add AI classes
The AI class exposes two static methods to get a list of best possible
moves for a player and to randomly select one in the list.
It uses the [BFS][bfs] algorithm to build a tree, where each node
represents a new movement. Each movement is ranked, and the best are
selected.

The rank of a movement is given by finding the players "goals", then
iterating over all it's pawns to find every position where the pawns can
move and rank them by their distance to the goals.